[ Seccion 8 ]
Refactorizacion de scroll_desing_screen. Se han extraido los widgets de main_content y de custom_background
