+++ diseno/lib/widgets/widgets.dart
@@ -1,3 +1,5 @@
+export 'package:diseno/widgets/custom_background.dart';
 export 'package:diseno/widgets/custom_icon.dart';
 export 'package:diseno/widgets/custom_text.dart';
 export 'package:diseno/widgets/custom_title.dart';
+export 'package:diseno/widgets/main_content.dart';
